feat: DB like -> ES 도입

+++ src/main/java/ounce/market/demo/product/dto/ProductRegisterCommand.java
@@ -3,7 +3,7 @@ package ounce.market.demo.product.dto;
 public record ProductRegisterCommand(
         String productCode,
         String name,
-        int basePrice,
+        Long basePrice,
         String description,
         String imageUrl,
         int initialStock
